LayoutTests/platform/mac-leopard/accessibility/table-sections-expected.txt

Example #1: Nested Stubs
1	a	b
1	e	f
1	c	d
1	c1	d1


1	a	b
1	c	d
1	c1	d1
1	c	d
1	c1	d1
1	c	d
1	c1	d1
1	e	f
--------------------------
THEAD, TFOOT table
--------------------------



--------------------------



--------------------------



--------------------------



--------------------------



--------------------------



--------------------------

--------------------------
Multi-TBODY table
--------------------------



--------------------------



--------------------------



--------------------------



--------------------------



--------------------------



--------------------------
